a StatefulSet without consecutive pods
Change the code to not assume that the ordinal has any meaning beyond
identity. An ordinal should never be re-used across pods with different
identity. A pod can be re-created with the same ordinal.

Assuming the presence of consecutive pods is not compatible
with fault-tolerant stateful services.

For example, during a failure we need to create a new pod with a new
identity and wait for data to replicate to it.
If the failed pod does not come online then we delete it.
This will leave a hole in the consecutive ordinals.
That failed pod should be permanently gone.
Filling its hole with a new pod can lead to a confusion of identity.